backend/requirements.txt
# AEGIS Backend — Python Dependencies
# Requires Python 3.9+

# FastAPI server (for full API)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Image processing
Pillow>=9.0.0
numpy>=1.24.0

# Neural model (transformers pipeline)
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0

# Minimum install (forensic ensemble only, no neural model)
# pip install Pillow numpy
